build: create copr Makefile
From now on, the maintainer must change the default version of the
software in the file resources/packaging/get_version. This script is
called by the copr Makefile to build the packages with the correct
version.

If a git directory is found in the current path, get_version will use
git-describe to set the version of the package. This implies to create
annotated tags when publishing a release.

The spec file is replaced by an input template file. Update the workflow
rpmbuild-action to make use of this new spec template. The Makefile also
uses the spec template.

To build the source RPM outside of Copr, use:

 $ make -f .copr/Makefile srpm outdir=/tmp/build/