feat: set sysconfig OPTIONS for chronyd
Add a new inventory variable time_chronyd_options to configure the value
of OPTIONS in /etc/sysconfig/chronyd. For full list of options, see
chronyd(8).

With this change, it is now possible to disable the control of the
system clock when running chronyd in containers, which is especially
useful when running CI in rootless containers.

Test this new feature with all previously defined hosts in molecule.
Create a new host 'client-no-start' which does not change the
/etc/sysconfig/chronyd and does not start the chronyd service.

Bump role time's version to 1.1.0 and update changelogs.